Introduce library versioning and notify about library changes
The library version number is stored in the admin table. It increases
after a bulk scan, a inotify callback or a spotify playlist change.
After increasing the version all listeners listening for the DATABASE
event are notified about the change.

A DB upgrade is done to store the initial library version.