[spotify] Remove "Spotify Saved" playlist
This special playlist was initially introduced due to the limited
ibspotify api to avoid purging tracks scanned from saved albums.
Since the move to the Spotify web api for scanning, it has no purpose.

Keeping it adds a measurable overhead during scan time (thousands of
inserts are done to playlistitems table).

A user can easily recreate this playlist with a custom smart playlist.